+++ plugin/index.d.ts
@@ -1,0 +1,2 @@
+export * from './joystick/joystick.d';
+export * from './keyboard/keyboard.d';