+++ plugin/changelog.md
@@ -1,6 +1,17 @@
-### [1.0.0](https://github.com/tencentyun/cloudgame-js-sdk/releases/tag/v1.1.3)@2022.02.21
+### [1.0.0](https://github.com/tencentyun/cloudgame-js-sdk/releases/tag/v1.1.3)@2022.03.23
 
 **新增**
 
-- 虚拟摇杆插件
+- 新增虚拟摇杆插件
 
+
+### [1.0.1](https://github.com/tencentyun/cloudgame-js-sdk/releases/tag/v1.1.4)@2022.05.16
+
+**新增**
+
+- 新增键盘插件
+
+**修复**
+
+- 摇杆插件在vue 环境构建无法操作问题
+- position 为0，造成操作无效问题